pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hf-mcp"
version = "0.1.0"
description = "Standalone Hack Forums MCP launch surface"
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["PyYAML>=6.0", "mcp>=1.0.0"]

[project.scripts]
hf-mcp = "hf_mcp.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
